(BSR)[BO] chore: move pro search into pro folder with dedicated blueprint
