+++ api/src/pcapi/routes/backoffice/templates/pro/get.html
@@ -4,6 +4,6 @@
   <p>{{ row.id }}</p>
   <p>{{ pro_type }}</p>
   <p>
-    <a href="{{ url_for('backoffice_web.search_pro') }}">Page de recherche (pro)</a>
+    <a href="{{ url_for('backoffice_web.pro.search_pro') }}">Page de recherche (pro)</a>
   </p>
 {% endblock page %}
